+++ tik_manager/SmNuke.py
@@ -31,7 +31,8 @@
 # -----------------------------------------------------------------------------
 
 import os
-import sys
+os.environ["FORCE_QT4"]="0"
+
 from SmUIRoot import MainUI as baseUI
 from SmRoot import RootManager
 
@@ -567,3 +568,6 @@ class MainUI(baseUI):
 
         # idk why this became necessary for nuke..
         self.category_tabWidget.setMaximumSize(QtCore.QSize(16777215, 30))
+
+        self.mIconPixmap = QtGui.QPixmap(os.path.join(self.manager._pathsDict["generalSettingsDir"], "icons", "iconNuke.png"))
+        self.managerIcon_label.setPixmap(self.mIconPixmap)