+++ tik_manager/SmPhotoshop.py
@@ -32,11 +32,11 @@
 import datetime
 import socket
 import sys, os
-import shutil
+# import shutil
 
 # Set the force pyqt environment variable to tell the other modulea not to use Qt.py module
-os.environ["FORCE_QT4"]="True"
-os.environ["PS_APP"]="True"
+os.environ["FORCE_QT4"]="1"
+os.environ["PS_APP"]="1"
 
 from PyQt4 import QtCore, Qt
 from PyQt4 import QtGui as QtWidgets
@@ -49,13 +49,14 @@ from SmUIRoot import MainUI as baseUI
 import _version
 # import subprocess
 
-# from win32com.client import Dispatch
-import comtypes.client as ct
+from win32com.client import Dispatch
+# import comtypes.client as ct
 
 # import pprint
 import logging
 
-psApp = ct.CreateObject('Photoshop.Application')
+# psApp = ct.CreateObject('Photoshop.Application')
+# psApp = "anan"
 
 __author__ = "Arda Kutlu"
 __copyright__ = "Copyright 2018, Scene Manager for Photoshop"
@@ -81,9 +82,136 @@ class PsManager(RootManager):
         self.exportFormats16Bit =["png", "jpg", "tif", "psd"]
         self.exportFormats32Bit =["exr", "tif", "hdr", "psd"]
 
-        self.textureTypes = ["Diffuse", "Opacity", "Reflection", "Subsurface", "Displacement", "Normal", "Cavity", "AO", "ID", "Mask", "Custom"]
         self.init_paths()
         self.init_database()
+        self.textureTypes = self._sceneManagerDefaults["exportTextureTypes"]
+        # self.psApp = ct.CreateObject('Photoshop.Application')
+        self.psApp = Dispatch('Photoshop.Application')
+
+    def init_paths(self):
+        """Overriden function"""
+        logger.debug("Func: init_paths")
+        _softwarePathsDict = self.getSoftwarePaths()
+        self._pathsDict["userSettingsDir"] = os.path.normpath(os.path.join(self.getUserDirectory(), _softwarePathsDict["userSettingsDir"]))
+        self._folderCheck(self._pathsDict["userSettingsDir"])
+
+        self._pathsDict["bookmarksFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smBookmarks.json"))
+        self._pathsDict["currentsFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smCurrents.json"))
+        self._pathsDict["projectsFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smProjects.json"))
+
+        self._pathsDict["commonFolderFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smCommonFolder.json"))
+
+        self._pathsDict["projectDir"] = self.getProjectDir()
+        self._pathsDict["sceneFile"] = ""
+
+        self._pathsDict["masterDir"] = os.path.normpath(os.path.join(self._pathsDict["projectDir"], "smDatabase"))
+        self._folderCheck(self._pathsDict["masterDir"])
+
+        self._pathsDict["databaseDir"] = os.path.normpath(os.path.join(self._pathsDict["masterDir"], _softwarePathsDict["databaseDir"]))
+        self._folderCheck(self._pathsDict["databaseDir"])
+
+        self._pathsDict["scenesDir"] = os.path.normpath(os.path.join(self._pathsDict["projectDir"], _softwarePathsDict["scenesDir"]))
+        self._folderCheck(self._pathsDict["scenesDir"])
+
+        self._pathsDict["projectSettingsFile"] = os.path.normpath(os.path.join(self._pathsDict["masterDir"], "projectSettings.json"))
+
+        self._pathsDict["subprojectsFile"] = os.path.normpath(os.path.join(self._pathsDict["masterDir"], "subPdata.json"))
+        self._pathsDict["categoriesFile"] = os.path.normpath(os.path.join(self._pathsDict["databaseDir"], _softwarePathsDict["categoriesFile"]))
+        self._pathsDict["previewsDir"] = os.path.normpath(os.path.join(self._pathsDict["projectDir"], "Playblasts", _softwarePathsDict["niceName"])) # dont change
+        self._pathsDict["pbSettingsFile"] = os.path.normpath(os.path.join(self._pathsDict["previewsDir"], _softwarePathsDict["pbSettingsFile"]))
+
+        self._pathsDict["generalSettingsDir"] = self._getCommonFolder()
+        if self._pathsDict["generalSettingsDir"] == -1:
+            self._exception(201, "Cannot Continue Without Common Database")
+            return -1
+
+        self._pathsDict["usersFile"] = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "sceneManagerUsers.json"))
+
+        self._pathsDict["softwareDatabase"] = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "softwareDatabase.json"))
+        self._pathsDict["sceneManagerDefaults"] = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "sceneManagerDefaults.json"))
+
+    def _checkCommonFolder(self, folder):
+        checkList = [os.path.join(folder, "sceneManagerDefaults.json"),
+                     os.path.join(folder, "sceneManagerUsers.json"),
+                     os.path.join(folder, "softwareDatabase.json")]
+        missingList = [os.path.basename(path) for path in checkList if not os.path.isfile(path)]
+        if len(missingList) > 0:
+
+            errorBox = QtWidgets.QMessageBox()
+            errorBox.setText("Common Database Folder missing some necessary files")
+            errorBox.setDetailedText("Following files are missing:\n %s" %pprint.pformat(missingList))
+            errorBox.exec_()
+            return False
+        else:
+            return True
+
+    def _getCommonFolder(self):
+        """Standalone Specific function - Different from software modules,
+        Standalone version needs to know where common files exist"""
+        if os.path.isfile(self._pathsDict["commonFolderFile"]):
+            commonFolder = self._loadJson(self._pathsDict["commonFolderFile"])
+            if commonFolder == -2:
+                return -2
+
+        else:
+            # in case there is no file found (Initial Run)
+            q = QtWidgets.QMessageBox()
+            q.setIcon(QtWidgets.QMessageBox.Question)
+            q.setText("Please define the Common Database Folder which contains general database files\n\nPress Ok to continue")
+            # q.setInformativeText(textInfo)
+            q.setWindowTitle("Select the Common Database Folder")
+            q.setDetailedText(
+                "Common Database Folder is the one which has the common modules for all software modules. Common Database Folder must include:\n"
+                "sceneManagerDefaults.json\n"
+                "sceneManagerUsers.json\n"
+                "softwareDatabase.json")
+            q.setStandardButtons(
+                QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Abort)
+
+            # q.button(QtWidgets.QMessageBox.Ok).setFixedHeight(30)
+            q.button(QtWidgets.QMessageBox.Ok).setFixedWidth(100)
+            #
+            # q.button(QtWidgets.QMessageBox.Abort).setFixedHeight(30)
+            q.button(QtWidgets.QMessageBox.Abort).setFixedWidth(100)
+            ret = q.exec_()
+            if ret == QtWidgets.QMessageBox.Ok:
+                commonFolder = self._defineCommonFolder()
+
+            elif ret == QtWidgets.QMessageBox.Abort:
+                return -1
+
+        return commonFolder
+
+    def _defineCommonFolder(self):
+        dlg = QtWidgets.QFileDialog.getExistingDirectory()
+        if dlg:
+            selectedDir = os.path.normpath(str(dlg))
+            if self._checkCommonFolder(selectedDir):
+                commonFolder = selectedDir
+                self._saveCommonFolder(commonFolder)
+
+                q = QtWidgets.QMessageBox()
+                q.setIcon(QtWidgets.QMessageBox.Information)
+                q.setText("Common Database Defined Successfully")
+                # q.setInformativeText(textInfo)
+                q.setWindowTitle("Success")
+                q.exec_()
+
+                return commonFolder
+            else:
+                return self._getCommonFolder()
+        else:
+            return self._getCommonFolder()
+            # self._getCommonFolder()
+
+    def _saveCommonFolder(self, data):
+        try:
+            self._dumpJson(data, self._pathsDict["commonFolderFile"])
+            msg = ""
+            return 0, msg
+        except:
+            msg = "Cannot save common folder file"
+            return -1, msg
 
     def getSoftwarePaths(self):
         """Overriden function"""
@@ -136,7 +264,7 @@ class PsManager(RootManager):
         # """This method must be overridden to return the full scene path ('' for unsaved) of current scene"""
         logger.debug("Func: getSceneFile")
         try:
-            activeDocument = psApp.Application.ActiveDocument
+            activeDocument = self.psApp.Application.ActiveDocument
             docName = activeDocument.name
             docPath = activeDocument.path
             return os.path.join(docPath, docName)
@@ -146,7 +274,7 @@ class PsManager(RootManager):
     def getFormatOptions(self):
         """returns the format options according to the bit depth of active document"""
         try:
-            activeDocument = psApp.Application.ActiveDocument
+            activeDocument = self.psApp.Application.ActiveDocument
             bitDepth = activeDocument.bitsPerChannel
         except:
             return []
@@ -222,11 +350,11 @@ class PsManager(RootManager):
         relSceneFile = os.path.relpath(sceneFile, start=projectPath)
 
 
-        openDocs = psApp.Application.Documents
+        openDocs = self.psApp.Application.Documents
         if openDocs.Count == 0:
-            activeDocument = psApp.Documents.Add(2048, 2048, 72)
+            activeDocument = self.psApp.Documents.Add(2048, 2048, 72)
         else:
-            activeDocument = psApp.Application.ActiveDocument
+            activeDocument = self.psApp.Application.ActiveDocument
 
         if sceneFormat == "psd":
             # PhotoshopSaveOptions=ct.CreateObject("Photoshop.PhotoshopSaveOptions")
@@ -236,26 +364,26 @@ class PsManager(RootManager):
             # PhotoshopSaveOptions.SpotColors = True
             # activeDocument.SaveAs(sceneFile, PhotoshopSaveOptions, False)
 
-            desc19 = ct.CreateObject("Photoshop.ActionDescriptor")
-            desc20 = ct.CreateObject("Photoshop.ActionDescriptor")
-            desc20.putBoolean(psApp.StringIDToTypeID('maximizeCompatibility'), True)
-            desc19.putObject(
-                psApp.CharIDToTypeID('As  '), psApp.CharIDToTypeID('Pht3'), desc20)
-            desc19.putPath(psApp.CharIDToTypeID('In  '), sceneFile)
-            desc19.putBoolean(psApp.CharIDToTypeID('LwCs'), True)
-            psApp.executeAction(psApp.CharIDToTypeID('save'), desc19, 3)
+            desc19 = Dispatch("Photoshop.ActionDescriptor")
+            desc20 = Dispatch("Photoshop.ActionDescriptor")
+            desc20.PutBoolean(self.psApp.StringIDToTypeID('maximizeCompatibility'), True)
+            desc19.PutObject(
+                self.psApp.CharIDToTypeID('As  '), self.psApp.CharIDToTypeID('Pht3'), desc20)
+            desc19.PutPath(self.psApp.CharIDToTypeID('In  '), sceneFile)
+            desc19.PutBoolean(self.psApp.CharIDToTypeID('LwCs'), True)
+            self.psApp.ExecuteAction(self.psApp.CharIDToTypeID('save'), desc19, 3)
 
         else:
 
-            desc19 = ct.CreateObject("Photoshop.ActionDescriptor")
-            desc20 = ct.CreateObject("Photoshop.ActionDescriptor")
-            desc20.putBoolean(psApp.StringIDToTypeID('maximizeCompatibility'), True)
+            desc19 = Dispatch("Photoshop.ActionDescriptor")
+            desc20 = Dispatch("Photoshop.ActionDescriptor")
+            desc20.PutBoolean(self.psApp.StringIDToTypeID('maximizeCompatibility'), True)
 
-            desc19.putObject(
-                psApp.CharIDToTypeID('As  '), psApp.CharIDToTypeID('Pht8'), desc20)
-            desc19.putPath(psApp.CharIDToTypeID('In  '), sceneFile)
-            desc19.putBoolean(psApp.CharIDToTypeID('LwCs'), True)
-            psApp.executeAction(psApp.CharIDToTypeID('save'), desc19, 3)
+            desc19.PutObject(
+                self.psApp.CharIDToTypeID('As  '), self.psApp.CharIDToTypeID('Pht8'), desc20)
+            desc19.PutPath(self.psApp.CharIDToTypeID('In  '), sceneFile)
+            desc19.PutBoolean(self.psApp.CharIDToTypeID('LwCs'), True)
+            self.psApp.ExecuteAction(self.psApp.CharIDToTypeID('save'), desc19, 3)
 
         thumbPath = self.createThumbnail(dbPath=jsonFile, versionInt=version)
 
@@ -266,7 +394,7 @@ class PsManager(RootManager):
         jsonInfo["ReferencedVersion"] = None
 
         jsonInfo["ID"] = "SmNukeV02_sceneFile"
-        jsonInfo["PSVersion"] = psApp.Version
+        jsonInfo["PSVersion"] = self.psApp.Version
         jsonInfo["Name"] = baseName
         jsonInfo["Path"] = os.path.relpath(shotPath, start=projectPath)
         jsonInfo["Category"] = categoryName
@@ -334,27 +462,27 @@ class PsManager(RootManager):
                 # PhotoshopSaveOptions.SpotColors = True
                 # activeDocument.SaveAs(sceneFile, PhotoshopSaveOptions, False)
 
-                desc19 = ct.CreateObject("Photoshop.ActionDescriptor")
-                desc20 = ct.CreateObject("Photoshop.ActionDescriptor")
-                desc20.putBoolean(psApp.StringIDToTypeID('maximizeCompatibility'), True)
+                desc19 = Dispatch("Photoshop.ActionDescriptor")
+                desc20 = Dispatch("Photoshop.ActionDescriptor")
+                desc20.PutBoolean(self.psApp.StringIDToTypeID('maximizeCompatibility'), True)
 
-                desc19.putObject(
-                    psApp.CharIDToTypeID('As  '), psApp.CharIDToTypeID('Pht3'), desc20)
-                desc19.putPath(psApp.CharIDToTypeID('In  '), sceneFile)
-                desc19.putBoolean(psApp.CharIDToTypeID('LwCs'), True)
-                psApp.executeAction(psApp.CharIDToTypeID('save'), desc19, 3)
+                desc19.PutObject(
+                    self.psApp.CharIDToTypeID('As  '), self.psApp.CharIDToTypeID('Pht3'), desc20)
+                desc19.PutPath(self.psApp.CharIDToTypeID('In  '), sceneFile)
+                desc19.PutBoolean(self.psApp.CharIDToTypeID('LwCs'), True)
+                self.psApp.ExecuteAction(self.psApp.CharIDToTypeID('save'), desc19, 3)
 
             else:
 
-                desc19 = ct.CreateObject("Photoshop.ActionDescriptor")
-                desc20 = ct.CreateObject("Photoshop.ActionDescriptor")
-                desc20.putBoolean(psApp.StringIDToTypeID('maximizeCompatibility'), True)
+                desc19 = Dispatch("Photoshop.ActionDescriptor")
+                desc20 = Dispatch("Photoshop.ActionDescriptor")
+                desc20.PutBoolean(self.psApp.StringIDToTypeID('maximizeCompatibility'), True)
 
-                desc19.putObject(
-                    psApp.CharIDToTypeID('As  '), psApp.CharIDToTypeID('Pht8'), desc20)
-                desc19.putPath(psApp.CharIDToTypeID('In  '), sceneFile)
-                desc19.putBoolean(psApp.CharIDToTypeID('LwCs'), True)
-                psApp.executeAction(psApp.CharIDToTypeID('save'), desc19, 3)
+                desc19.PutObject(
+                    self.psApp.CharIDToTypeID('As  '), self.psApp.CharIDToTypeID('Pht8'), desc20)
+                desc19.PutPath(self.psApp.CharIDToTypeID('In  '), sceneFile)
+                desc19.PutBoolean(self.psApp.CharIDToTypeID('LwCs'), True)
+                self.psApp.ExecuteAction(self.psApp.CharIDToTypeID('save'), desc19, 3)
 
             thumbPath = self.createThumbnail(dbPath=jsonFile, versionInt=currentVersion)
 
@@ -439,81 +567,93 @@ class PsManager(RootManager):
 
         filePath = os.path.join(exportFolderPath, fileName)
 
-        activeDocument = psApp.Application.ActiveDocument
+        if os.path.isfile(filePath):
+            if not self._query(("The following file will be overwritten if you continue:\n %s" %fileName), "Choose Yes to overwrite file and continue"):
+                return
+            else:
+                pass
+
+        activeDocument = self.psApp.Application.ActiveDocument
         if extension == "jpg":
-            saveOPT=ct.CreateObject("Photoshop.JPEGSaveOptions")
+            saveOPT=Dispatch("Photoshop.JPEGSaveOptions")
             saveOPT.EmbedColorProfile = True
             saveOPT.FormatOptions = 1 # => psStandardBaseline
             saveOPT.Matte = 1 # => No Matte
             saveOPT.Quality = 12
             activeDocument.SaveAs(filePath, saveOPT, True)
         if extension == "png":
-            saveOPT=ct.CreateObject("Photoshop.PNGSaveOptions")
+            saveOPT=Dispatch("Photoshop.PNGSaveOptions")
             activeDocument.SaveAs(filePath, saveOPT, True)
         if extension == "bmp":
-            saveOPT=ct.CreateObject("Photoshop.BMPSaveOptions")
+            saveOPT=Dispatch("Photoshop.BMPSaveOptions")
             activeDocument.SaveAs(filePath, saveOPT, True)
         if extension == "tga":
-            saveOPT=ct.CreateObject("Photoshop.TargaSaveOptions")
+            saveOPT=Dispatch("Photoshop.TargaSaveOptions")
             saveOPT.Resolution=32
             saveOPT.AlphaChannels=True
             saveOPT.RLECompression=True
             activeDocument.SaveAs(filePath, saveOPT, True)
         if extension == "psd":
-            saveOPT=ct.CreateObject("Photoshop.PhotoshopSaveOptions")
+            saveOPT=Dispatch("Photoshop.PhotoshopSaveOptions")
             saveOPT.AlphaChannels = True
             saveOPT.Annotations = True
             saveOPT.Layers = True
             saveOPT.SpotColors = True
             activeDocument.SaveAs(filePath, saveOPT, True)
         if extension == "tif":
-            saveOPT=ct.CreateObject("Photoshop.TiffSaveOptions")
+            saveOPT=Dispatch("Photoshop.TiffSaveOptions")
             saveOPT.AlphaChannels = True
             saveOPT.EmbedColorProfile = True
             saveOPT.Layers = False
             activeDocument.SaveAs(filePath, saveOPT, True)
         if extension == "exr":
-            # desc19 = ct.CreateObject("Photoshop.ActionDescriptor")
-            # desc20 = ct.CreateObject("Photoshop.ActionDescriptor")
-            # idBtDp = app.CharIDToTypeID(str('BtDp'))
-            # desc20.putInteger(idBtDp, 16)
-            # idCmpr = app.CharIDToTypeID('Cmpr')
-            # desc20.putInteger(idCmpr, 1)
-            # idAChn = app.CharIDToTypeID('AChn')
-            # desc20.putInteger(idAChn, 0)
-            # desc19.putObject(app.CharIDToTypeID('As  '), app.CharIDToTypeID('EXRf'), desc20)
-            # desc19.putPath(app.CharIDToTypeID('In  '), str(filePath))
-            # app.executeAction(app.CharIDToTypeID('save'), desc19, 3)
-            idsave = psApp.CharIDToTypeID("save")
-            desc182 = ct.CreateObject("Photoshop.ActionDescriptor")
-            idAs = psApp.CharIDToTypeID("As  ")
-            desc183 = ct.CreateObject("Photoshop.ActionDescriptor")
-            idBtDp = psApp.CharIDToTypeID("BtDp")
-            desc183.putInteger(idBtDp, 16);
-            idCmpr = psApp.CharIDToTypeID("Cmpr")
-            desc183.putInteger(idCmpr, 1)
-            idAChn = psApp.CharIDToTypeID("AChn")
-            desc183.putInteger(idAChn, 0)
-            idEXRf = psApp.CharIDToTypeID("EXRf")
-            desc182.putObject(idAs, idEXRf, desc183)
-            idIn = psApp.CharIDToTypeID("In  ")
-            desc182.putPath(idIn, (filePath))
-            idDocI = psApp.CharIDToTypeID("DocI")
-            desc182.putInteger(idDocI, 340)
-            idCpy = psApp.CharIDToTypeID("Cpy ")
-            desc182.putBoolean(idCpy, True)
-            idsaveStage = psApp.StringIDToTypeID("saveStage")
-            idsaveStageType = psApp.StringIDToTypeID("saveStageType")
-            idsaveSucceeded = psApp.StringIDToTypeID("saveSucceeded")
-            desc182.putEnumerated(idsaveStage, idsaveStageType, idsaveSucceeded)
-            psApp.executeAction(idsave, desc182, 3)
+            idsave = self.psApp.CharIDToTypeID("save")
+            desc182 = Dispatch("Photoshop.ActionDescriptor")
+            idAs = self.psApp.CharIDToTypeID("As  ")
+            desc183 = Dispatch("Photoshop.ActionDescriptor")
+            idBtDp = self.psApp.CharIDToTypeID("BtDp")
+            desc183.PutInteger(idBtDp, 16);
+            idCmpr = self.psApp.CharIDToTypeID("Cmpr")
+            desc183.PutInteger(idCmpr, 1)
+            idAChn = self.psApp.CharIDToTypeID("AChn")
+            desc183.PutInteger(idAChn, 0)
+            idEXRf = self.psApp.CharIDToTypeID("EXRf")
+            desc182.PutObject(idAs, idEXRf, desc183)
+            idIn = self.psApp.CharIDToTypeID("In  ")
+            desc182.PutPath(idIn, (filePath))
+            idDocI = self.psApp.CharIDToTypeID("DocI")
+            desc182.PutInteger(idDocI, 340)
+            idCpy = self.psApp.CharIDToTypeID("Cpy ")
+            desc182.PutBoolean(idCpy, True)
+            idsaveStage = self.psApp.StringIDToTypeID("saveStage")
+            idsaveStageType = self.psApp.StringIDToTypeID("saveStageType")
+            idsaveSucceeded = self.psApp.StringIDToTypeID("saveSucceeded")
+            desc182.PutEnumerated(idsaveStage, idsaveStageType, idsaveSucceeded)
+            self.psApp.ExecuteAction(idsave, desc182, 3)
 
         if extension == "hdr":
-            pass
+            idsave = self.psApp.CharIDToTypeID("save")
+            desc419 = Dispatch("Photoshop.ActionDescriptor")
+            idAs = self.psApp.CharIDToTypeID("As  ")
+            desc419.PutString(idAs, """Radiance""")
+            idIn = self.psApp.CharIDToTypeID("In  ")
+            desc419.PutPath(idIn, (filePath))
+            idDocI = self.psApp.CharIDToTypeID("DocI")
+            desc419.PutInteger(idDocI, 333)
+            idCpy = self.psApp.CharIDToTypeID("Cpy ")
+            desc419.PutBoolean(idCpy, True)
+            idsaveStage = self.psApp.StringIDToTypeID("saveStage")
+            idsaveStageType = self.psApp.StringIDToTypeID("saveStageType")
+            idsaveSucceeded = self.psApp.StringIDToTypeID("saveSucceeded")
+            desc419.PutEnumerated(idsaveStage, idsaveStageType, idsaveSucceeded)
+            self.psApp.ExecuteAction(idsave, desc419, 3)
 
-
-
-        pass
+        success = os.path.isfile(filePath)
+        if success:
+            return filePath
+        else:
+            self._exception(202, "Unknown Error")
+            return False
 
 
     def loadBaseScene(self, force=False):
@@ -522,7 +662,7 @@ class PsManager(RootManager):
         relSceneFile = self._currentSceneInfo["Versions"][self._currentVersionIndex-1]["RelativePath"]
         absSceneFile = os.path.join(self.projectDir, relSceneFile)
         if os.path.isfile(absSceneFile):
-            psApp.Open(absSceneFile)
+            self.psApp.Open(absSceneFile)
             return 0
         else:
             msg = "File in Scene Manager database doesnt exist"
@@ -557,8 +697,8 @@ class PsManager(RootManager):
         # create a thumbnail using playblast
         thumbDir = os.path.split(thumbPath)[0]
         if os.path.exists(thumbDir):
-            psApp.Preferences.RulerUnits = 1
-            activeDocument = psApp.Application.ActiveDocument
+            self.psApp.Preferences.RulerUnits = 1
+            activeDocument = self.psApp.Application.ActiveDocument
             dupDocument = activeDocument.Duplicate("thumbnailCopy", True)
             dupDocument.bitsPerChannel = 8
             oWidth = 221
@@ -619,10 +759,31 @@ class PsManager(RootManager):
             if categoriesData == -2:
                 return -2
         else:
-            # categoriesData = self._sceneManagerDefaults["defaultCategories"]
-            categoriesData = ["Concept", "Storyboard", "Texture", "Other"]
+            categoriesData = self._sceneManagerDefaults["defaultPSCategories"]
+            # categoriesData = ["Concept", "Storyboard", "Texture", "Other"]
             self._dumpJson(categoriesData, self._pathsDict["categoriesFile"])
         return categoriesData
+
+    def _query(self, header, msg):
+        q = QtWidgets.QMessageBox()
+        q.setIcon(QtWidgets.QMessageBox.Question)
+        q.setText(header)
+        q.setInformativeText(msg)
+        q.setWindowTitle("Are you sure?")
+        q.setStandardButtons(
+            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
+        q.button(QtWidgets.QMessageBox.Yes).setFixedHeight(30)
+        q.button(QtWidgets.QMessageBox.Yes).setFixedWidth(100)
+        q.button(QtWidgets.QMessageBox.Cancel).setFixedHeight(30)
+        q.button(QtWidgets.QMessageBox.Cancel).setFixedWidth(100)
+
+        ret = q.exec_()
+        if ret == QtWidgets.QMessageBox.Yes:
+            return True
+        else:
+            return False
+
+
 
     def _exception(self, code, msg):
         """OVERRIDEN"""
@@ -683,6 +844,10 @@ class MainUI(baseUI):
 
         self.export_pushButton.setVisible(True)
         self.export_pushButton.clicked.connect(self.exportSourceUI)
+
+
+        self.mIconPixmap = QtWidgets.QPixmap(os.path.join(self.manager._pathsDict["generalSettingsDir"], "icons", "iconPS.png"))
+        self.managerIcon_label.setPixmap(self.mIconPixmap)
         #
         # self.baseScene_label.setVisible(False)
         # self.baseScene_lineEdit.setVisible(False)
@@ -821,7 +986,48 @@ class MainUI(baseUI):
             textureType = self.customType_lineEdit.text() if self.type_comboBox.currentText() == "Custom" else self.type_comboBox.currentText()
             asNextRevision = self.incremental_checkBox.isChecked()
             revisionNumber = version_spinBox.value()
-            self.manager.exportSourceimage(extension=extension, textureType=textureType, asNextRevision=asNextRevision, revisionNumber=revisionNumber)
+            exportedFilePath = self.manager.exportSourceimage(extension=extension, textureType=textureType, asNextRevision=asNextRevision, revisionNumber=revisionNumber)
+            if exportedFilePath:
+                successUI("Success", destPath=exportedFilePath)
+            else:
+                successUI("Failure")
+            self.exportTexture_Dialog.close()
+
+        def successUI(status, color="white", destPath=None):
+
+            self.msgDialog = QtWidgets.QDialog(parent=self)
+            self.msgDialog.setModal(True)
+            self.msgDialog.setWindowTitle("Export Successfull")
+            self.msgDialog.resize(300, 120)
+            layoutMain = QtWidgets.QVBoxLayout()
+            self.msgDialog.setLayout(layoutMain)
+
+            infoHeader = QtWidgets.QLabel(status)
+
+            infoHeader.setStyleSheet(""
+                                     "border: 18px solid black;"
+                                     "background-color: black;"
+                                     "font-size: 14px;"
+                                     "color: {0}"
+                                     "".format(color))
+
+            layoutMain.addWidget(infoHeader)
+            layoutH = QtWidgets.QHBoxLayout()
+            layoutMain.addLayout(layoutH)
+
+            openFile = QtWidgets.QPushButton("Open File")
+            layoutH.addWidget(openFile)
+            openFile.clicked.connect(lambda x: self.manager.showInExplorer(destPath))
+            showInExplorer = QtWidgets.QPushButton("Show in Explorer")
+            layoutH.addWidget(showInExplorer)
+            okButton = QtWidgets.QPushButton("OK")
+            layoutH.addWidget(okButton)
+
+            showInExplorer.clicked.connect(lambda x: self.manager.showInExplorer(os.path.split(destPath)[0]))
+
+            okButton.clicked.connect(self.msgDialog.close)
+
+            self.msgDialog.show()
 
         # SIGNALS
         # -------
@@ -832,11 +1038,14 @@ class MainUI(baseUI):
         buttonC.clicked.connect(self.exportTexture_Dialog.reject)
 
         buttonE.clicked.connect(exportCommand)
-        buttonE.clicked.connect(self.exportTexture_Dialog.accept)
+        # buttonE.clicked.connect(self.exportTexture_Dialog.accept)
 
 
 
         self.exportTexture_Dialog.show()
+
+
+
 
 if __name__ == '__main__':
     selfLoc = os.path.dirname(os.path.abspath(__file__))