+++ tik_manager/assetLibrary.py
@@ -40,7 +40,8 @@ import _version
 import datetime
 
 
-FORCE_QT4 = bool(os.getenv("FORCE_QT4"))
+# FORCE_QT4 = bool(os.getenv("FORCE_QT4"))
+FORCE_QT4 = bool(int(os.environ["FORCE_QT4"]))
 
 # Enabele FORCE_QT4 for compiling with pyinstaller
 # FORCE_QT4 = True